client/requirements.txt
azure-core==1.31.0
azure-cosmos==4.7.0
azure-storage-blob==12.23.1
